agentic_intent/requirements.txt
fastmcp
duckduckgo-search
python-dotenv
langgraph
langchain-openai
langchain-core
langchain-groq
openai
mcp
anyio